# openapi.yml
openapi: 3.1.0
info:
  title: CrackCS API
  version: 5.0.0
  description: 세션 인증, 콘텐츠 운영, 근거 검색과 구조화 평가 API
servers:
  - url: /
...
paths:
...
  /api/questions/{questionId}/answers:
    parameters:
      - $ref: '#/components/parameters/QuestionIdPath'
    post:
      tags: [Answers]
      operationId: submitAnswer
      summary: USER 답변 저장과 평가 접수
      description: 같은 회원의 동일 키·동일 문제·동일 원문은 기존 결과 반환. 다른 payload는 409. 재답변은 새 키 사용.
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            format: uuid
            pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$'
        - name: X-CSRF-TOKEN
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  minLength: 1
                  maxLength: 10000
                  description: 공백만 있는 값 거부. 원문 공백 보존.
      responses:
        '202':
          description: 저장된 답변과 현재 평가 상태. 재요청 시 이미 완료된 상태 가능.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnswerResponse'
        '400':
          description: 필수 헤더 누락 또는 입력 오류
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/QuestionNotFound'
        '409':
          description: 같은 키의 다른 문제 또는 답변 원문
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/members/me/answers:
    get:
      tags: [Answers]
      operationId: findMyAnswers
      summary: USER 본인의 답변 이력
      description: submittedAt DESC, answerId DESC 고정 정렬
      parameters:
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/SizeQuery'
      responses:
        '200':
          description: 답변 이력 페이지
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnswerPageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/answers/{answerId}:
    parameters:
      - $ref: '#/components/parameters/AnswerIdPath'
    get:
      tags: [Answers]
      operationId: findMyAnswer
      summary: USER 본인의 답변 상세
      responses:
        '200':
          description: 폐기된 문제의 본문도 이력에서 제공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnswerResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: 답변 없음 또는 다른 회원 소유
  /api/answers/{answerId}/evaluation:
    parameters:
      - $ref: '#/components/parameters/AnswerIdPath'
    get:
      tags: [Answers]
      operationId: findMyEvaluation
      summary: USER 본인의 평가 상태와 결과
      description: 평가 상태, 구조화 피드백과 실제 사용 근거를 제공.
      responses:
        '200':
          description: 평가 진행·완료·실패 상태
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: 답변 없음 또는 다른 회원 소유
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: CRACKCS_SESSION
  parameters:
    AnswerIdPath:
      name: answerId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    QuestionIdPath:
      name: questionId
      in: path
      required: true
      description: 문제 ID
      schema:
        type: integer
        format: int64
        minimum: 1
...
    PageQuery:
      name: page
      in: query
      required: false
      description: 0부터 시작하는 페이지 번호
      schema:
        type: integer
        minimum: 0
        default: 0
    SizeQuery:
      name: size
      in: query
      required: false
      description: 페이지 크기
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
...
    QuestionNotFound:
      description: 해당 ID의 문제가 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            code: QUESTION_NOT_FOUND
            message: 'Question not found: 999'
            fieldErrors: []
            requestId: 1e85b909-2114-47be-a1c3-1fa47a4a7235
    Unauthorized:
      description: 인증이 필요함
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Forbidden:
      description: 인증됐지만 역할 또는 소유권이 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
...
  schemas:
    AnswerResponse:
      type: object
      required: [answerId, evaluationId, questionId, questionContent, content, submittedAt, evaluation]
      properties:
        answerId:
          type: integer
          format: int64
        evaluationId:
          type: integer
          format: int64
        questionId:
          type: integer
          format: int64
        questionContent:
          type: string
        content:
          type: string
        submittedAt:
          type: string
          description: 서버 LocalDateTime, ISO-8601 offset 없는 시각
        evaluation:
          $ref: '#/components/schemas/EvaluationResponse'
    AnswerPageResponse:
      type: object
      required: [content, page, size, totalElements, totalPages]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/AnswerResponse'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
    EvaluationResponse:
      type: object
      required: [status, verdict, score, feedback, failureReason, concepts, strengths, omissions, misconceptions, evidence]
      properties:
        status:
          type: string
          enum: [EVALUATING, PROCESSING, EVALUATED, NEEDS_REVIEW, FAILED]
        verdict:
          type: [string, 'null']
          enum: [CORRECT, PARTIALLY_CORRECT, INCORRECT, NEEDS_REVIEW, null]
        score:
          type: [integer, 'null']
          enum: [100, 50, 0, null]
        feedback:
          type: [string, 'null']
        failureReason:
          type: [string, 'null']
          enum: [PROVIDER_ERROR, PROVIDER_TIMEOUT, INVALID_RESULT, PROVIDER_UNAVAILABLE, EVIDENCE_NOT_FOUND, EVIDENCE_CONFLICT, MONTHLY_BUDGET_EXCEEDED, null]
        concepts:
          type: array
          items:
            type: object
            required: [conceptId, conceptName, verdict, score, feedback]
            properties:
              conceptId:
                type: integer
                format: int64
              conceptName:
                type: string
              verdict:
                type: string
                enum: [CORRECT, PARTIALLY_CORRECT, INCORRECT, NEEDS_REVIEW]
              score:
                type: [integer, 'null']
                enum: [100, 50, 0, null]
              feedback:
                type: string
        strengths:
          type: array
          items: { type: string }
        omissions:
          type: array
          items: { type: string }
        misconceptions:
          type: array
          items: { type: string }
        evidence:
          type: array
          items:
            $ref: '#/components/schemas/EvaluationEvidenceResponse'
...
    EvaluationEvidenceResponse:
      type: object
      required: [chunkId, documentTitle, documentVersion, startOffset, endOffset, content]
      properties:
        chunkId: { type: integer, format: int64 }
        documentTitle: { type: string }
        documentVersion: { type: integer, minimum: 1 }
        startOffset: { type: integer, minimum: 0 }
        endOffset: { type: integer, minimum: 1 }
        content: { type: string }
...
    ApiErrorResponse:
      type: object
      required:
        - code
        - message
        - fieldErrors
        - requestId
      properties:
        code:
          type: string
        message:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorResponse'
        requestId:
          type: string
          format: uuid
    FieldErrorResponse:
      type: object
      required:
        - field
        - reason
      properties:
        field:
          type: string
        reason:
          type: string
...
security:
  - sessionCookie: []